Add ICS tests to the interchaintest suite.
Includes tests for ics versions 3.3 and 4; also has the code for a test
with neutron and stride, but neutron isn't launching yet, so I disabled
the test.

The ICS tests work by launching a consumer chain, upgrading the
provider, checking that the consumer still works (i.e. stake delegations
from the provider are reflected in the consumer, and IBC transfers work)
then launching a new consumer chain and checking that _it_ works.