[WIP] Add LSM tests.
Includes WIP tests for LSM happy paths, as we had in v15, plus a new
test for tokenizing vested delegations. The latter just creates a
vesting account, waits for half the vesting time to elapse, and tries to
tokenize a) the full vesting amount, which fails, and b) half the
vesting amount, i.e. what's vested, which suceeds in v16.

Also includes a basic epoch test that will set blocksPerEpoch to 20 and
ensure that it in fact takes 20 blocks to update validator powers in a
consumer chain.